+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change log
 
+## 0.2.0 (release date: 8.06.2017)
+
+* Added `eslint-plugin-goog` and `eslint-plugin-interfaced` to peer dependencies
+* Added plugins `goog` and `interfaced` to `plugins` section
+* Added rules from plugin `goog`
+* Added rules from plugin `interfaced`
+* Allow methods in `no-empty-function` (for extending by `interfaced/no-empty-method`)
+* Removed `@virtual` from `valid-jsdoc` blacklist
+
 ## 0.1.0 (release date: 29.03.2017)
 
 * Initial release